Separate create_stackfull_coroutine functions
